+++ app/src/app/components/stix/list-property/list-property.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { StixObject } from 'src/app/classes/stix/stix-object';
 
 @Component({
@@ -6,12 +6,11 @@ import { StixObject } from 'src/app/classes/stix/stix-object';
   templateUrl: './list-property.component.html',
   styleUrls: ['./list-property.component.scss']
 })
-export class ListPropertyComponent implements OnInit {
+export class ListPropertyComponent {
     @Input() public config: ListPropertyConfig;
 
-    constructor() { }
-
-    ngOnInit(): void {
+    constructor() {
+        // intentionally left blank
     }
 
 }