+++ app/src/app/views/stix/collection/collection-view/collection-view.component.ts
@@ -22,6 +22,8 @@ import { environment } from "../../../../../environments/environment";
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { logger } from "../../../../util/logger";
 import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
+import { MatDialog } from '@angular/material/dialog';
+import { CollectionUpdateDialogComponent } from 'src/app/components/collection-update-dialog/collection-update-dialog.component';
 import { Campaign } from 'src/app/classes/stix/campaign';
 
 type changeCategory = "additions" | "changes" | "minor_changes" | "revocations" | "deprecations";
@@ -46,8 +48,24 @@ export class CollectionViewComponent extends StixViewPage implements OnInit {
     public validating: boolean = false;
     public validationData: ValidationData = null;
 
+    // pluralize attackType for text display
+    public plural = {
+        "technique": "techniques",
+        "tactic": "tactics",
+        "group": "groups",
+        "software": "software",
+        "mitigation": "mitigations",
+        "matrix": "matrices",
+        "data-source": "data sources",
+        "data-component": "data components"
+    }
+
     public get collectionDownloadURL() {
         return `${environment.integrations.rest_api.url}/collection-bundles/?collectionId=${this.collection.stixID}&collectionModified=${this.collection.modified.toISOString()}`
+    }
+
+    public get hasTechniques() {
+        return this.collectionChanges.technique.object_count > 0;
     }
 
     public stagedData: VersionReference[] = [];
@@ -86,7 +104,8 @@ export class CollectionViewComponent extends StixViewPage implements OnInit {
                 private restApiConnector: RestApiConnectorService,
                 private snackbar: MatSnackBar,
                 private editor: EditorService,
-                authenticationService: AuthenticationService) {
+                public dialog: MatDialog,
+                authenticationService: AuthenticationService,) {
         super(authenticationService);
     }
 
@@ -181,6 +200,8 @@ export class CollectionViewComponent extends StixViewPage implements OnInit {
                     if (object.created_by_ref) identities.add(object.created_by_ref);
                     if (object.modified_by_ref) identities.add(object.modified_by_ref);
                     for (let marking_ref of object.object_marking_refs) marking_defs.add(marking_ref);
+
+                    if (object.type == "marking-definition") continue;
 
                     this.stagedData.push(new VersionReference({
                         "object_ref": object.stixID,
@@ -538,6 +559,55 @@ export class CollectionViewComponent extends StixViewPage implements OnInit {
 
     public format(attackType: string): string {
         return attackType.replace(/_/g, ' ');
+    }
+
+    /**
+     * Determine if the auto add objects button is disabled
+     * for a given attack type
+     * @param {string} attackType the type of object to add
+     * @returns {boolean} false, if objects of this type can be added
+     * (button is not disabled), true otherwise (button is disabled)
+     */
+    public autoAddDisabled(attackType: string): boolean {
+        if (attackType == 'tactic') return !this.hasTechniques;
+        return true; // only tactics are currently supported for this feature
+    }
+
+    /**
+     * Open the update dialog to automatically find and add
+     * related objects to the collection
+     * @param {string} attackType the type of objects to add to the collection
+     */
+    public addObjectsToCollection(attackType: string): void {
+        let prompt = this.dialog.open(CollectionUpdateDialogComponent, {
+            data: {
+                collectionChanges: this.collectionChanges,
+                potentialChanges: this.potentialChanges,
+                attackType: attackType
+            },
+            maxHeight: "75vh"
+        })
+        let subscription = prompt.afterClosed().subscribe({
+            next: result => {
+                if (result) {
+                    // reinitialize stix lists
+                    this.editingReloadToggle = false;
+                    setTimeout(() => this.editingReloadToggle = true);
+                }
+            },
+            complete: () => { subscription.unsubscribe(); }
+        });
+    }
+
+    /**
+     * Get the display text list of object types that will be used
+     * to identify objects to automatically add to the collection
+     * @param attackType the type of objects to add to the collection
+     * @returns {string} display text of relevant objects
+     */
+    public relevantObjects(attackType: string): string {
+        if (attackType == 'tactic') return 'techniques';
+        return 'objects'; // only tactics are currently supported for auto-add
     }
 
     ngOnInit() {